modifica dello stato del device tramite uuid

+++ src/main/java/it/customers/h2db/springboot/repositry/CustomerRepository.java
@@ -1,11 +1,9 @@
 package it.customers.h2db.springboot.repositry;
 
 import it.customers.h2db.springboot.models.Customer;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CustomerRepository extends CrudRepository<Customer, Integer> {
+public interface CustomerRepository extends CrudRepository<Customer, Long> {
 
-/*  @Query("SELECT c FROM Customer c WHERE c.codiceFiscale = ?1")*/
   Customer findByCodiceFiscale(String codiceFiscale);
 }
